pyproject.toml
[tool.poetry]
name = "drf-spectacular-test"
version = "0.1.0"
description = ""
authors = ["JngMkk <kjm0817ss@gmail.com>"]
readme = "README.md"

[tool.black]
target-version = ['py310']
line-length = 100

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
drf-spectacular = "^0.27.2"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
